requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
colorama==0.4.3
Django==3.0.6
django-rest-auth==0.9.5
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
